+++ packages/atomic/src/components/facets/atomic-category-facet/atomic-category-facet.tsx
@@ -382,7 +382,7 @@ export class AtomicCategoryFacet
         hasActiveValues={this.facetState.hasActiveValues}
       >
         {this.facetSearch?.render()}
-        <div class="mt-1 lg:text-sm">
+        <div class="mt-1 text-lg lg:text-base">
           {this.allCategoriesButton}
           <div>{this.parents}</div>
           <div class={this.parents.length ? 'pl-9' : 'pl-0'}>