+++ packages/headless/src/utils/errors.ts
@@ -7,3 +7,11 @@ export class ExpiredTokenError extends Error {
     this.message = 'The token being used to perform the request is expired.';
   }
 }
+
+export class DisconnectedError extends Error {
+  constructor() {
+    super();
+    this.name = 'Disconnected';
+    this.message = 'Client is not connected to the internet.';
+  }
+}